syntaxes: use the CSS captures
This looks mostly fine but there are two issues. I had to make a change
to the comment block so it would not interfere. Note the `\S` appended.
The other thing is that we are not colorizing the `###`.

Regardless I think this basic support would let me use Visual Studio
Code more for Imba. Also these can be resolved better when we revamp the
language server support.